federated remote mediator cleanup
