+++ feature/feed/src/main/java/social/firefly/feed/FeedModule.kt
@@ -25,16 +25,5 @@ val feedModule = module {
         analyticsModule,
     )
 
-    singleOf(::HomeTimelineRemoteMediator)
-    single {
-        LocalTimelineRemoteMediator(
-            get(),
-        )
-    }
-    single {
-        FederatedTimelineRemoteMediator(
-            get(),
-        )
-    }
     viewModelOf(::FeedViewModel)
 }
